fix: avoid primitive values
Primitive values can't be used as generics. This is a problem when
trying to serialize, for example, a Vec<u16>, which would be a
java.util.ArrayList<int>. However, the primitive type int cannot be used
as a generic of ArrayList. Therefore, for now, we will use the
non-primitive object versions of each primitive type e.g. Integer
instead of int. This leaves room for optimisation, as in some cases we
_can_ use the primitive value just fine. This can be picked up as a
future improvement if necessary.